feat: ini next sentry
