--- apps/app/next.config.js
@@ -1,29 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  experimental: {
-    instrumentationHook: true,
-    // From: https://github.com/vercel/next.js/issues/42641
-    outputFileTracingExcludes: {
-      "*": [
-        "./node_modules/@swc/core-linux-x64-gnu",
-        "./node_modules/@swc/core-linux-x64-musl",
-        "./node_modules/esbuild-linux-64/bin",
-        "./node_modules/webpack/lib",
-        "./node_modules/rollup",
-        "./node_modules/terser",
-      ],
-    },
-    serverActions: true,
-  },
-  outputFileTracing: true,
-  transpilePackages: ["@lightdotso/trpc", "@lightdotso/ui"],
-  webpack: config => {
-    config.externals.push("async_hooks", "pino-pretty", "lokijs", "encoding");
-    config.resolve.fallback = { fs: false, net: false, tls: false };
-
-    return config;
-  },
-};
-
-module.exports = nextConfig;
