+++ apps/app/package.json
@@ -33,6 +33,8 @@
     "@opentelemetry/sdk-trace-node": "^1.15.2",
     "@opentelemetry/semantic-conventions": "^1.15.2",
     "@prisma/client": "^5.2.0",
+    "@sentry/integrations": "^7.70.0",
+    "@sentry/nextjs": "^7.70.0",
     "@tanstack/react-query": "^5.0.0-beta.9",
     "@tanstack/react-query-devtools": "^5.0.0-beta.9",
     "@tanstack/react-query-next-experimental": "^5.0.0-beta.9",
