+++ apps/app/turbo.json
@@ -9,9 +9,13 @@
         "NEXTAUTH_URL",
         "NEXTAUTH_SECRET",
         "NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID",
+        "NEXT_PUBLIC_SENTRY_DSN",
         "NEXT_PUBLIC_SIGNIN_MESSAGE",
+        "NEXT_PUBLIC_VERCEL_ENV",
         "NEXT_PUBLIC_VERCEL_URL",
-        "NEXT_RUNTIME"
+        "NEXT_RUNTIME",
+        "SENTRY_DSN",
+        "VERCEL_ENV"
       ],
       "inputs": ["app/**", "pages/**", "src/**", "next.config.js"],
       "outputs": ["dist/**", ".next/**", "!.next/cache/**"]
